pyproject.toml
[project]
name = "riset-baileys-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
]
